Work around buggy wavefront parsing
Wavefront gets really upset if we send spans with "duplicate" tags.

We can send the span "kind" either as a "span.kind" tag or as part of
the "kind" field in the top-level object.  If we send it in both, then
Wavefront sees this as "duplicate" tags.  This causes problems in their
backend which then caused them to block all our spans.